% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Box Plots}
\usage{
box_plot(data, elements, group, xlegend = "")
}
\arguments{
\item{data}{Dataframe}

\item{elements}{Select the elements to plot as vector.}

\item{group}{Select the grouping of the data set as character variable. This needs to be the
name of a column in the data frame.}

\item{xlegend}{Label the x-axis.}
}
\value{

}
\description{
Each element is a single plot of grouped boxplots.
}
