% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-icp.R
\name{.replace_lod_values}
\alias{.replace_lod_values}
\title{Replaces <LOD, <0.000 and N/A values in ICP-MS with zeroes}
\usage{
.replace_lod_values(x)
}
\arguments{
\item{x}{The column in the data frame to be processed.}
}
\value{
The processed column.
}
\description{
Replaces <LOD, <0.000 and N/A values in ICP-MS with zeroes
}
\examples{
.replace_lod_values(df$x)

}
\seealso{
\code{\link[=process_icp]{process_icp()}}
}
