% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_values_checker.R
\name{double_values_checker}
\alias{double_values_checker}
\title{Helper function for the Laser Mapping function. Checks the potential legend
breaks if some breaks occur twice. This might happen since values are rounded
to look nicer.}
\usage{
double_values_checker(
  breaks_fun = small_breaks_list[[1]],
  small_scale = small_breaks_list[[2]]
)
}
\arguments{
\item{breaks_fun}{The values for breaks defined stored in a list defined by the Laser Mapping
function.}

\item{small_scale}{Boolean value that decides if a small scale is necassary.}
}
\description{
Helper function for the Laser Mapping function. Checks the potential legend
breaks if some breaks occur twice. This might happen since values are rounded
to look nicer.
}
