% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_value_plot.R
\name{eigen_value_plot}
\alias{eigen_value_plot}
\title{Eigen Value Plot}
\usage{
eigen_value_plot(pcobj, sel_pc = c(1, 2), ...)
}
\arguments{
\item{pcobj}{PCA object from stats::pcomp.}

\item{sel_pc}{Select two principal components to plot.}

\item{...}{size of the labels and ratio of the plot}
}
\value{

}
\description{
Plot the Eigen Values and their respective loadings from a PCA. Simple plot
that doesn't include the data points.
}
