% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dare.R
\name{geom_dare}
\alias{geom_dare}
\title{geom_dare}
\usage{
geom_dare(...)
}
\arguments{
\item{...}{Pass arguments on to geom_text and geom_line}
}
\description{
geoms to draw fields defined by Dare et al. (2014)"
}
