% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pca_arrows.R
\name{geom_pca_arrows}
\alias{geom_pca_arrows}
\title{Eigen Value plot geom for tidymodel PCA}
\usage{
geom_pca_arrows(
  pca_rec,
  pc_x = 1,
  pc_y = 2,
  labels = TRUE,
  labels_fill = "white",
  labels_color = "#0A537D",
  labels_size = 0.25,
  label_font_size = 3,
  label_rect_padding = 0.25,
  labels_font_size = 3,
  labels_rect_padding = 0.25,
  ...
)
}
\arguments{
\item{pca_rec}{The PCA recipe, a tidymodel object.}

\item{pc_x}{PC shown on x-axis.}

\item{pc_y}{PC shown on y-axis.}

\item{labels}{If TRUE labels of the arrows are plotted.}

\item{labels_fill}{Fill color of the labels}

\item{labels_color}{Color of the labels text}

\item{labels_size}{Size of the labels}

\item{label_font_size}{Size of the font in the labels (used by geom_label_repel).}

\item{label_rect_padding}{Size of the padding around the labels (used by geom_label_repel).}

\item{labels_font_size}{Deprecated alias for \code{label_font_size}.}

\item{labels_rect_padding}{Deprecated alias for \code{label_rect_padding}.}

\item{...}{Arguments passed on to geom_label_repel, which is drawing the arrows.}
}
\value{
A list of ggplot2 layer objects.
}
\description{
Eigen Value plot geom for tidymodel PCA
}
\examples{
\dontrun{
ggplot2::ggplot() + geom_pca_arrows(my_recipe)
}
}
