% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Histograms for all selected columns}
\usage{
hist_plot(data, x_axis, bins)
}
\arguments{
\item{bins}{}
}
\value{
ggplot objects
}
\description{
Histograms for all selected columns
}
