% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgt_class_D_B.R
\name{mgt_class_D_B}
\alias{mgt_class_D_B}
\title{Magnetite Classification Template after Dupuis&Beaudoin}
\usage{
mgt_class_D_B()
}
\value{
ggplot object
}
\description{
Magnetite Classification Template after Dupuis&Beaudoin
}
\examples{

}
