% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxides_to_elements.R
\name{oxides_to_elements}
\alias{oxides_to_elements}
\title{Oxides to Elements}
\usage{
oxides_to_elements(data)
}
\arguments{
\item{data}{Dataframe as tibble}
}
\value{
A tibble containing the elements.
}
\description{
Convert oxides found in the whole rock geochemical data sets to elements.
}
