% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA score plot}
\usage{
pca_plot(pca_rec, pc_x = 1, pc_y = 2, data = NA, ...)
}
\arguments{
\item{pca_rec}{The PCA recipe, a tidymodel object.}

\item{pc_x}{PC shown on x-axis.}

\item{pc_y}{PC shown on y-axis.}

\item{data}{Optional data frame of pre-baked PCA scores. If \code{NA} (default) scores are
extracted from \code{pca_rec} via \code{\link[recipes:bake]{recipes::bake()}}.}

\item{...}{arguments passed on to labs.}
}
\value{
A ggplot object.
}
\description{
PCA score plot
}
\examples{
\dontrun{
pca_plot(my_recipe, pc_x = 1, pc_y = 2)
}
}
