% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gibbs.R
\name{plot_base_gibbs}
\alias{plot_base_gibbs}
\title{Plots the base Gibbs diagram}
\usage{
plot_base_gibbs()
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the base Gibbs diagram
}
\examples{
plot_base_gibbs()

}
\seealso{
\code{\link[=plot_base_major_anions]{plot_base_major_anions()}}

\code{\link[=plot_base_ll]{plot_base_ll()}}

\code{\link[=plot_base_major_cations]{plot_base_major_cations()}}

\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=plot_base_water_maturity]{plot_base_water_maturity()}}
}
