% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-langelier-ludwig.R
\name{plot_base_ll}
\alias{plot_base_ll}
\title{Plots the base Langelier-Ludwig diagram}
\usage{
plot_base_ll()
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the base Langelier-Ludwig diagram
}
\examples{
plot_base_ll()

}
\seealso{
\code{\link[=plot_base_gibbs]{plot_base_gibbs()}}

\code{\link[=plot_base_gigg_anions]{plot_base_gigg_anions()}}

\code{\link[=plot_base_major_cations]{plot_base_major_cations()}}

\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=plot_base_water_maturity]{plot_base_water_maturity()}}
}
