% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-major-anions.R
\name{plot_base_major_anions}
\alias{plot_base_major_anions}
\title{Plots the base ternary diagram for major anions (Giggenbach)}
\usage{
plot_base_major_anions()
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the base ternary diagram for major anions (Giggenbach)
}
\examples{
plot_base_major_anions()

}
\seealso{
\code{\link[=plot_base_gibbs]{plot_base_gibbs()}}

\code{\link[=plot_base_ll]{plot_base_ll()}}

\code{\link[=plot_base_major_cations]{plot_base_major_cations()}}

\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=plot_base_water_maturity]{plot_base_water_maturity()}}
}
