% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-major-cations.R
\name{plot_base_major_cations}
\alias{plot_base_major_cations}
\title{Plots the base ternary diagram for major cations}
\usage{
plot_base_major_cations()
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the base ternary diagram for major cations
}
\examples{
plot_base_major_cations()

}
\seealso{
\code{\link[=plot_base_gibbs]{plot_base_gibbs()}}

\code{\link[=plot_base_major_anions]{plot_base_major_anions()}}

\code{\link[=plot_base_ll]{plot_base_ll()}}

\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=plot_base_water_maturity]{plot_base_water_maturity()}}
}
