% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-piper.R
\name{plot_base_piper}
\alias{plot_base_piper}
\title{Plots the base Piper plot}
\usage{
plot_base_piper()
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the base Piper plot
}
\examples{
plot_base_piper()

}
\references{
{
\insertRef{piper1944_GraphicProcedureGeochemical}{geochem}
}
}
\seealso{
\code{\link[=transform_data_piper]{transform_data_piper()}}
}
