% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metals.R
\name{plot_metals}
\alias{plot_metals}
\title{Plots the absolute and relative metal concentration as bar plots}
\usage{
plot_metals(df)
}
\arguments{
\item{df}{Data frame containing sample, concentration and element columns}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots the absolute and relative metal concentration as bar plots
}
\seealso{
\code{\link[=process_icp]{process_icp()}}
}
