% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-ic.R
\name{process_ic}
\alias{process_ic}
\title{Processes IC results for further geochemistry plots:
\itemize{
\item Calculates miliequivalents for the given species
\item Adds coordinate transformations for the Langelier-Ludwig diagram
\item Adds percentage transformations for the Piper plot
}}
\usage{
process_ic(df)
}
\arguments{
\item{df}{Data frame containing the following columns:
\itemize{
\item Anions: alk_tot, cl, so4, br, no3, po4
\item Cations: na, k, ca, mg, nh4, li
}}
}
\value{
A data.frame object.
}
\description{
Processes IC results for further geochemistry plots:
\itemize{
\item Calculates miliequivalents for the given species
\item Adds coordinate transformations for the Langelier-Ludwig diagram
\item Adds percentage transformations for the Piper plot
}
}
\examples{
df = process_ic(df)

}
\seealso{
\code{\link[=plot_base_gibbs]{plot_base_gibbs()}}

\code{\link[=plot_base_gigg_anions]{plot_base_gigg_anions()}}

\code{\link[=plot_base_ll]{plot_base_ll()}}

\code{\link[=plot_base_major_cations]{plot_base_major_cations()}}

\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=plot_base_water_maturity]{plot_base_water_maturity()}}
}
