% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-icp.R
\name{process_icp}
\alias{process_icp}
\title{Processes ICP-MS as given by the instrument}
\usage{
process_icp(filepath, blank_name)
}
\arguments{
\item{filepath}{String indicating the location of the raw data file.}

\item{blank_name}{String indicating the name of the blank in the file.}
}
\value{
A data.frame object.
}
\description{
Processes ICP-MS as given by the instrument
}
\examples{
process_icp(filepath="icp-ms.csv", blank_name="blank")

}
\seealso{
\code{\link[=.replace_lod_values]{.replace_lod_values()}}

\code{\link[=select_icp_auto]{select_icp_auto()}}

\code{\link[=plot_metals]{plot_metals()}}
}
