% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-Plots}
\usage{
qq_plot(
  data,
  elements,
  group,
  color_scale = ggplot2::scale_color_viridis_d(),
  theme = ggplot2::theme_dark(),
  print = TRUE
)
}
\arguments{
\item{data}{Datatable}

\item{elements}{List with the elements that should be plotted.}

\item{group}{Name of the group used for coloring the plot.}

\item{color_scale}{Specify the colorscale. Virdis is the standard color scale.}

\item{theme}{Specify the theme. ggplot2::theme_dark() is the stadard theme.}

\item{print}{If TRUE qq plots are printed directly if FALSE the qq-plots are
saved to a list, which is returned.}
}
\description{
Creates qq-plots for all elements specified and colors the plot
according to group.
}
