% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Number_Imputer.R
\name{Random_Number_Imputer}
\alias{Random_Number_Imputer}
\title{Impute Random Numbers}
\usage{
Random_Number_Imputer(data, columns, split_symbol = "<", digits = 6)
}
\value{
data
}
\description{
Impute a random number below the detection limit.
}
