% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{Rounds up nicely, used in the Laser Mapping function.}
\usage{
round_up(x, nice = c(seq(1, 10, 0.1)))
}
\arguments{
\item{x}{Vetor with values that should be rounded up.}

\item{nice}{Defines to which values x should be rounded up.}
}
\value{

}
\description{
Rounds up nicely, used in the Laser Mapping function.
}
\examples{

}
