% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-icp.R
\name{select_icp_auto}
\alias{select_icp_auto}
\title{Automatically selects the processed ICP-MS data according to defined checks}
\usage{
select_icp_auto(df, blank_name)
}
\arguments{
\item{df}{Data frame as given by \code{\link[=process_icp]{process_icp()}} function}

\item{blank_name}{String indicating the name of the blank in the file.}
}
\value{
A data.frame object.
}
\description{
Automatically selects the processed ICP-MS data according to defined checks
}
\examples{
select_icp_auto(df=df, blank_name="blank")

}
\seealso{
\code{\link[=process_icp]{process_icp()}}
}
