% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_breaks_definer.R
\name{small_breaks_definer}
\alias{small_breaks_definer}
\title{Helper functions that defines legend breaks for the Laser Maps.
Only used for linear scales.}
\usage{
small_breaks_definer(
  min_scale,
  mid_scale,
  max_scale,
  small_scale,
  min_ppm,
  max_ppm,
  smallest_scale_I,
  smallest_scale_II,
  smallest_scale_III
)
}
\arguments{
\item{min_scale}{Minimum of the scale}

\item{mid_scale}{Middle of the scale}

\item{max_scale}{Maximum of the scale}

\item{small_scale}{Boolean Value defining if a small scale is necassary}

\item{min_ppm}{Minimum value for the column}

\item{max_ppm}{Maximum value for the column}

\item{smallest_scale_I}{Defines where the breaks in the legend occur}

\item{smallest_scale_II}{Defines where the breaks in the legend occur}

\item{smallest_scale_III}{Defines where the breaks in the legend occur}
}
\value{

}
\description{
Helper functions that defines legend breaks for the Laser Maps.
Only used for linear scales.
}
