% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas_diagram.R
\name{tas_diagram}
\alias{tas_diagram}
\title{Total Alkali Silica diagram}
\usage{
tas_diagram(fields = TRUE, labels = TRUE)
}
\arguments{
\item{fields}{If TRUE the TAS fields are drawn.}

\item{labels}{If TRUE the names for the respective field are drawn.}
}
\description{
Template for the popular Total Alkali Silica (TAS) diagram after Le Maitre (1989)
The function should be used as a geom in a ggplot.
}
\examples{
ggplot2::ggplot() +
    geochem::tas_diagram()
}
