% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_glab}
\alias{theme_glab}
\title{Giovannelli lab theme (https://github.com/giovannellilab/GLab_ggplot2_theme)}
\usage{
theme_glab(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/180,
  base_rect_size = base_size/180
)
}
\arguments{
\item{base_size}{Default text size}

\item{base_family}{Default text font}

\item{base_line_size}{Default line size}

\item{base_rect_size}{Default rect size (not used)}
}
\value{
The Giovannelli lab theme.
}
\description{
Giovannelli lab theme (https://github.com/giovannellilab/GLab_ggplot2_theme)
}
