% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-piper.R
\name{transform_data_piper}
\alias{transform_data_piper}
\title{Transforms input concentrations to match the coordinates in the Piper plot.
Adapted from the following sources:
\itemize{
\item https://gist.github.com/johnDorian/5561272
\item https://github.com/markolipka/ggplot_Piper/blob/master/ggplot_Piper.R
}}
\usage{
transform_data_piper(Mg, Ca, Cl, SO4, id)
}
\value{
A data.frame object.
}
\description{
Transforms input concentrations to match the coordinates in the Piper plot.
Adapted from the following sources:
\itemize{
\item https://gist.github.com/johnDorian/5561272
\item https://github.com/markolipka/ggplot_Piper/blob/master/ggplot_Piper.R
}
}
\examples{
df = transform_data_piper(Mg, Ca, Cl, SO4, id)

}
\references{
{
\insertRef{piper1944_GraphicProcedureGeochemical}{geochem}
}
}
\seealso{
\code{\link[=plot_base_piper]{plot_base_piper()}}

\code{\link[=process_ic]{process_ic()}}
}
