% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping_element.R
\name{clipping_element}
\alias{clipping_element}
\title{Plotly Laser Map for selected element}
\usage{
clipping_element(elmnt, data)
}
\arguments{
\item{elmnt}{The element that will be clipped.}

\item{data}{The laser map data set.}
}
\value{

}
\description{
Shows a plotly map of the selected element to determine the cut off value
for that paritcular element. The function doesn;t do the actual clipping.
The Clipping has to be done manually with filter().
}
