% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_quantiles.R
\name{filter_quantiles}
\alias{filter_quantiles}
\title{filter_quantiles}
\usage{
filter_quantiles(data, .cols, probs = seq(0, 1, 0.05), quantile_position = 20)
}
\arguments{
\item{data}{Provide data frame in tibble format.}

\item{.cols}{Specify numeric columns that should be screened for NAs.}

\item{probs}{Sequence that determines step size for quantiles.}

\item{quantile_position}{Position in the sequence determined by probs. Values within quantiles higher than
quantile_position will be replaced by NA}
}
\value{
data
}
\description{
Replaces outliers with NAs.
}
