% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pca_arrows.R
\name{geom_pca_arrows}
\alias{geom_pca_arrows}
\title{Eigen Value plot geom for tidymodel PCA}
\usage{
geom_pca_arrows(pca_rec, pc_x = 1, pc_y = 2, labels = TRUE)
}
\arguments{
\item{pca_rec}{The PCA recipe, a tidymodel object.}

\item{pc_x}{PC shown on x-axis.}

\item{pc_y}{PC shown on y-axis.}

\item{labels}{If TRUE labels of the arrows are plotted.}
}
\value{
ggplot object
}
\description{
Eigen Value plot geom for tidymodel PCA
}
