% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser_map.R
\name{laser_map}
\alias{laser_map}
\title{Laser Maps Plot}
\usage{
laser_map(
  data,
  selected_elements,
  Log_Trans = FALSE,
  pca_rec = NA,
  sel_pc = NA,
  unit_title = "(ppm)",
  fontsize = 14,
  font = "serif",
  labels = scales::label_scientific(),
  ...
)
}
\arguments{
\item{data}{Laser map data frame. If PCA are plotted the PC columns must be merged with
the laser map data frame.}

\item{selected_elements}{Vector containing the selected elements and PC's will be plotted.}

\item{Log_Trans}{Data set containing a column called "Log" which indicates if a map should be
log transformed. If not specified all elements are log transformed.}

\item{sel_pc}{Selecting the principal components from the pcobj to calculate the explained
variance shown by the laser map.}

\item{unit_title}{Specify the unit of the laser map for the plot title as a string. Has no
effect on the values.}

\item{fontsize}{Fontsize used in the plot.}

\item{font}{Font type, options are "serif", "sans" and "mono".}

\item{labels}{Argument for the scale notation. See ?scales:: for more information.}

\item{...}{Options for viridis color scale. Includes "A", "B", "C", "D".
See ?ggplot2::scale_fill_viridis for more information.}

\item{pcobj}{PCA object created by stats::prcomp.}
}
\value{

}
\description{
Function that creates laser maps of selected elements. The function creates
a list containing the individual laser plots. The list can than used to
plot all the maps at once via ggpubr::ggarrange or individual plots can
be plotted by selecting single elements out of the list.
}
\examples{

}
