% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser_map2.R
\name{laser_map2}
\alias{laser_map2}
\title{Plot for Elemental Map}
\usage{
laser_map2(
  data,
  columns,
  option = "turbo",
  trans = "log",
  breaks = c(10^(-4:6)),
  labels = scales::label_scientific(),
  plot_label_start = "b",
  plot_label = letters,
  unit = "[ppm]",
  family = "serif",
  pca_rec = NA,
  option_Temp = "D"
)
}
\arguments{
\item{option_Temp}{}
}
\description{
Plot for Elemental Map
}
