% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laser_map2.R
\name{laser_map2}
\alias{laser_map2}
\title{Plot for Elemental Map}
\usage{
laser_map2(
  data,
  columns,
  option = "turbo",
  trans = "log",
  breaks = c(10^(-3:5)),
  labels = scales::label_scientific(),
  plot_label_start = "B",
  plot_label = LETTERS,
  unit = "[ppm]",
  pca_rec = NA
)
}
\arguments{
\item{pca_rec}{}
}
\value{

}
\description{
Plot for Elemental Map
}
\examples{

}
