% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Title}
\usage{
pca_plot(pca_rec, pc_x = 1, pc_y = 2, ...)
}
\arguments{
\item{pca_rec}{The PCA recipe, a tidymodel object.}

\item{pc_x}{PC shown on x-axis.}

\item{pc_y}{PC shown on y-axis.}

\item{...}{arguments passed on to labs.}
}
\value{
ggplot object
}
\description{
Title
}
