% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_map.R
\name{raster_map}
\alias{raster_map}
\title{Raster Map}
\usage{
raster_map(data, x, y, n = 100)
}
\arguments{
\item{data}{Provide dataframe with x and y coordinates.}

\item{x}{Column mapped to x coordinates}

\item{y}{Column mapped to y coordinates}

\item{n}{Number of data squares the old dataframe is divided into}
}
\value{
data
}
\description{
Computes a square raster with n tiles from two columns.
}
