% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmod_df_wrangler.R
\name{xmod_df_wrangler}
\alias{xmod_df_wrangler}
\title{Wrangles the XMOD data frames into a plotable format.}
\usage{
xmod_df_wrangler(
  directory,
  xls_common_pattern,
  csv_common_pattern,
  csv_list_to_plot = FALSE
)
}
\arguments{
\item{directory}{Directory where the XMOD files are located.}

\item{xls_common_pattern}{Pattern for the modal data .xls files}

\item{csv_common_pattern}{Pattern of the xmod files containing the coordinates .csv files}

\item{csv_list_to_plot}{Can use the geochem::xmod_samples_list function as argument.}
}
\description{
!! Warning this functions overwrites the original data frames!!
Loads the data frames if the wt\% value is missing. Adds the wt\% for
}
