% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmod_samples_list.R
\name{xmod_samples_list}
\alias{xmod_samples_list}
\title{Create list of XMOD files to plot}
\usage{
xmod_samples_list(file_pattern, directory)
}
\arguments{
\item{file_pattern}{String with the common name pattern of all csvs in the directory that should be plotted.}

\item{directory}{Location of the directory which contains the XMOD csv's.}
}
\value{

}
\description{
The function is loading all the the data frames in the specified directory that end with
.csv in a list. The list is used in the following function to load the
data frames into the current environment.
}
\examples{
xmod_samples_list('XMOD_Sample', './/XMOD//Samples')
}
